+++ bloom_nofos/nofos/views.py
@@ -1,3 +1,4 @@
+import csv
 import io
 import os
 
@@ -8,9 +9,10 @@ from constance import config
 from django.conf import settings
 from django.contrib import messages
 from django.contrib.admin.views.decorators import staff_member_required
+from django.contrib.auth.decorators import user_passes_test
 from django.db import transaction
 from django.db.models import F
-from django.http import HttpResponse, HttpResponseBadRequest, HttpResponseServerError
+from django.http import HttpResponse, HttpResponseBadRequest
 from django.shortcuts import get_object_or_404, redirect, render
 from django.urls import reverse_lazy
 from django.utils import dateformat, timezone
@@ -617,3 +619,22 @@ def insert_order_space_view(request, section_id):
 
     context = {"form": form, "title": "Insert Order Space", "section": section}
     return render(request, "admin/insert_order_space.html", context)
+
+
+@user_passes_test(lambda u: u.is_superuser)
+def export_nofo_links(request, nofo_id):
+    # Create the HttpResponse object with the appropriate CSV header.
+    response = HttpResponse(content_type="text/csv")
+    response["Content-Disposition"] = f'attachment; filename="nofo_{nofo_id}_links.csv"'
+
+    writer = csv.writer(response)
+
+    writer.writerow(["url", "section_name", "nofo_number"])
+
+    nofo = get_object_or_404(Nofo, id=nofo_id)
+
+    urls = find_external_links(nofo)
+    for url in urls:
+        writer.writerow([url.get("url"), url.get("section").name, nofo.number])
+
+    return response