+++ CHANGELOG.md
@@ -8,6 +8,8 @@ The format is based on Keep a Changelog, and this project adheres to Semantic Ve
 
 ### Added
 
+- Add admin-only route to export all external links from a NOFO
+
 ### Changed
 
 - More style maps added based on an early version of HHS-2026-ACL-AOD-DDUC